Default tcp route's TTL to 120 if not present
- change TTL type from uint16 to *int
- When a route with negative TTL value is created, the error
  returned is about not being able to render JSON. Instead a meaningful
  error should be returned

[#119118969]

+++ handlers/fakes/fake_validator.go
@@ -27,12 +27,12 @@ type FakeRouteValidator struct {
 	validateDeleteReturns struct {
 		result1 *routing_api.Error
 	}
-	ValidateCreateTcpRouteMappingStub        func(tcpRouteMappings []models.TcpRouteMapping, routerGroups models.RouterGroups, maxTTL uint16) *routing_api.Error
+	ValidateCreateTcpRouteMappingStub        func(tcpRouteMappings []models.TcpRouteMapping, routerGroups models.RouterGroups, maxTTL int) *routing_api.Error
 	validateCreateTcpRouteMappingMutex       sync.RWMutex
 	validateCreateTcpRouteMappingArgsForCall []struct {
 		tcpRouteMappings []models.TcpRouteMapping
 		routerGroups     models.RouterGroups
-		maxTTL           uint16
+		maxTTL           int
 	}
 	validateCreateTcpRouteMappingReturns struct {
 		result1 *routing_api.Error
@@ -112,12 +112,12 @@ func (fake *FakeRouteValidator) ValidateDeleteReturns(result1 *routing_api.Error
 	}{result1}
 }
 
-func (fake *FakeRouteValidator) ValidateCreateTcpRouteMapping(tcpRouteMappings []models.TcpRouteMapping, routerGroups models.RouterGroups, maxTTL uint16) *routing_api.Error {
+func (fake *FakeRouteValidator) ValidateCreateTcpRouteMapping(tcpRouteMappings []models.TcpRouteMapping, routerGroups models.RouterGroups, maxTTL int) *routing_api.Error {
 	fake.validateCreateTcpRouteMappingMutex.Lock()
 	fake.validateCreateTcpRouteMappingArgsForCall = append(fake.validateCreateTcpRouteMappingArgsForCall, struct {
 		tcpRouteMappings []models.TcpRouteMapping
 		routerGroups     models.RouterGroups
-		maxTTL           uint16
+		maxTTL           int
 	}{tcpRouteMappings, routerGroups, maxTTL})
 	fake.validateCreateTcpRouteMappingMutex.Unlock()
 	if fake.ValidateCreateTcpRouteMappingStub != nil {
@@ -133,7 +133,7 @@ func (fake *FakeRouteValidator) ValidateCreateTcpRouteMappingCallCount() int {
 	return len(fake.validateCreateTcpRouteMappingArgsForCall)
 }
 
-func (fake *FakeRouteValidator) ValidateCreateTcpRouteMappingArgsForCall(i int) ([]models.TcpRouteMapping, models.RouterGroups, uint16) {
+func (fake *FakeRouteValidator) ValidateCreateTcpRouteMappingArgsForCall(i int) ([]models.TcpRouteMapping, models.RouterGroups, int) {
 	fake.validateCreateTcpRouteMappingMutex.RLock()
 	defer fake.validateCreateTcpRouteMappingMutex.RUnlock()
 	return fake.validateCreateTcpRouteMappingArgsForCall[i].tcpRouteMappings, fake.validateCreateTcpRouteMappingArgsForCall[i].routerGroups, fake.validateCreateTcpRouteMappingArgsForCall[i].maxTTL
